pyproject.toml
[tool.poetry]
name = "cashkollab"
version = "0.0.1"
description = ""
authors = [
    "Lukas Becker <störte@ströte.wtf>",
    "David Bauer <hausprojekt@debauer.de>",
]
license = "MIT"
readme = "README.md"


packages = [
    { include = "hellcash", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.0.5"
django-rest-framework = "^0.1.0"
django-filter = "^22.1"
django-crispy-forms = "^1.14.0"
django-tables2 = "^2.4.1"
python-dateutil = "^2.8.2"
"backports.zoneinfo" = {version = "^0.2.1", optional = true }
overrides = "^6.1.0"
django-bootstrap5 = "^21.3"
pendulum = "^2.1.2"
pytest-mock = "^3.8.2"
pandas = "^1.4.3"
gitpython = "^3.1.40"


[tool.poetry.dev-dependencies]
black = "^22.3.0"
mypy = "0.961"
pylint = "2.13.8"
pyupgrade = "2.32.1"
isort = "5.10.1"
pytest = "^7.1.2"
django-stubs = "^1.12.0"
types-python-dateutil = "^2.8.18"
ruff = "^0.1.6"

[build-system]
requires = ["poetry_core>=1.0.0", "setuptools"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.pylint.master]
extension-pkg-whitelist = ["numpy", "cv2"]
jobs = 0
ignore = "migrations"

[tool.pylint.basic]
good-names = [
    "i", # iterator
    "_", "__", "___", # unneeded variable
    "T", # generic type
    "id",
    "pk"
]

[tool.pylint.message_control]
disable = [
    "missing-module-docstring", "missing-class-docstring", "missing-function-docstring",
    "line-too-long",
    "too-few-public-methods",
    "fixme",
    "duplicate-code",
    "too-many-ancestors",
]
enable = [
    "useless-suppression"
]

[tool.pylint.design]
max-attributes = 10

[tool.mypy]
python_version = "3.10"

junit_xml = "xunit-reports/xunit-result-mypy.xml"

# output style configuration
show_column_numbers = true
show_error_codes = true
pretty = true

# additional warnings
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_no_return = true

no_implicit_optional = true
warn_unreachable = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
# disallow_any_explicit = true
# disallow_any_generics = true
disallow_untyped_calls = true

ignore_missing_imports = true

plugins = ["mypy_django_plugin.main"]
exclude = "migrations"

[tool.django-stubs]
django_settings_module = "cashkollab.settings"